LayoutTests/imported/w3c/web-platform-tests/intersection-observer/clip-path-animation-expected.txt


FAIL clip-path-animation assert_approx_equals: entries[1].intersectionRect.left expected 0 +/- 0 but got 8
